extract PostmanRequestGenerator
